# Use faster loop iter (despite clippy's advice)